change int16_t to int8_t

+++ app/src/main/cpp/visualizer/execute_fft.c
@@ -207,7 +207,7 @@ int init_visualizer(filter_sys_t *p_filter)
     p_filter->data_size = 0;
     p_filter->nb_samples = 0;
     p_filter->out_samples = FFT_BUFFER_SIZE;
-    p_filter->output = (int16_t *) (malloc(p_filter->out_samples * sizeof(int16_t)));
+    p_filter->output = (int8_t *) (malloc(p_filter->out_samples * sizeof(int8_t)));
     return 0;
 }
